Modifications to the intro
The "flow" stalls at in the when the concept of file states (the figure) is introduced. I have added a sentence trying to "smooth" the intro of the file state figure.

Also added git reset HEAD .ignore git, since it is necessary for being able to list the same git status as in the example.