Update 06-branches.md
Let's see if this works and updates the proposed changes in the patch-2. 
I have modified the code to implement the things discussed here
https://github.com/coderefinery/git-intro/pull/277#issuecomment-729910546

The update to the [git-branch-2.svg](https://coderefinery.github.io/git-intro/img/gitink/git-branch-2.svg) visual remains still open.
The HEAD needs to point to l1 commit not to e2 commit.