link to PDF version; closes #429
This won't work very well for forks so what we should do instead
is that we link to the PDF of the current repo.

The reason why I did not do this is that the PDF does not exist
yet locally and thus local build would otherwise fail.

We can fix this later.